New post type "discussion" rendering added. Splited postdetail into question and discussioın pages. Updated backend to return post type for feed views
